+++ src/background/handlers/onResponseStarted.ts
@@ -1,5 +1,5 @@
 import { WebRequest } from "webextension-polyfill";
-import findChannelFromVideoWeaverUrl from "../../common/ts/findChannelFromVideoWeaverUrl";
+import { findChannelFromVideoWeaverUrl } from "../../common/ts/findChannel";
 import getHostFromUrl from "../../common/ts/getHostFromUrl";
 import getProxyInfoFromUrl from "../../common/ts/getProxyInfoFromUrl";
 import isChromium from "../../common/ts/isChromium";
