+++ src/background/handlers/onTabCreated.ts
@@ -1,8 +1,8 @@
 import { Tabs } from "webextension-polyfill";
 import getHostFromUrl from "../../common/ts/getHostFromUrl";
 import isChromium from "../../common/ts/isChromium";
+import { updateProxySettings } from "../../common/ts/proxySettings";
 import { twitchTvHostRegex } from "../../common/ts/regexes";
-import updateProxySettings from "../../common/ts/updateProxySettings";
 import store from "../../store";
 
 export default function onTabCreated(tab: Tabs.Tab): void {
