--- src/common/ts/clearProxySettings.ts
@@ -1,5 +1,0 @@
-export default function clearProxySettings() {
-  chrome.proxy.settings.clear({ scope: "regular" }, function () {
-    console.log("⚙️ Proxy settings cleared");
-  });
-}
